fix(warning): use explicit casts to silence sign-conversion and sign-compare warnings
- taihe_native_object.cpp ConvertPropertyToAni: cast operationIndication
  (masked to 0..255) directly to int32_t instead of unsigned int, matching
  the declared type; value unchanged
- nearlink_sle_datatransfer_service.cpp ReceivedData: compare Read result
  (int) against static_cast<int>(pLen); pLen is pre-checked to
  0 < pLen < SOCKET_BUFFER_SIZE (51200), so the cast is lossless

+++ services/service/src/datatransfer/nearlink_sle_datatransfer_service.cpp
@@ -533,7 +533,7 @@ bool SleDataTransferService::ReceivedData(std::shared_ptr<InputStream> inputStre
         (void)memset_s(pBuf, sizeof(pBuf), 0, sizeof(pBuf));
         int res = inputStream->Read(pBuf, sizeof(pBuf));
         NL_CHECK_RETURN_RET(res != 0, false, "fd disconnected err");
-        NL_CHECK_RETURN_RET(res == pLen, false, "data len err");
+        NL_CHECK_RETURN_RET(res == static_cast<int>(pLen), false, "data len err");
         std::shared_ptr<DataTransferDataParams> result = std::make_shared<DataTransferDataParams>();
         NearlinkDataTransferDataParams::DeserializeData(pBuf, pLen, *result);
         NL_CHECK_RETURN_RET(result->port_ == portId, false, "port not match");
